requirements.txt
feedparser==6.0.11
requests-html==0.10.0
pydantic==2.10.6
argparse
bs4==0.0.2
EbookLib==0.18
uvicorn==0.34.0
fastapi==0.115.8
